feat: add support to spanish on uploads
As we are using Jupiter to upload spanish videos from Digital House
it would be cool to support spanish language on new uploads.

This commit introduces a new select component on each upload where
the user can choose the video language so the transcription and
the description will be generated in that language.